🔧 rework : ExerciseSetController

+++ backend/backend_CalisthenX/src/main/java/org/example/backend_calisthenx/services/ExerciseSetService.java
@@ -1,5 +1,6 @@
 package org.example.backend_calisthenx.services;
 
+import org.example.backend_calisthenx.exceptions.ResourceNotFoundException;
 import org.example.backend_calisthenx.models.ExerciseSet;
 import org.example.backend_calisthenx.repositories.ExerciseSetRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,15 +28,29 @@ public class ExerciseSetService {
     }
 
     public ExerciseSet updateExerciseSet(ExerciseSet exerciseSet) {
-        return exerciseSetRepository.save(exerciseSet); // La méthode save permet également de mettre à jour.
+        ExerciseSet updated = exerciseSetRepository.findById(exerciseSet.getId())
+                .orElseThrow(()-> new ResourceNotFoundException("Exercise Set not found"));
+
+        updated.setExerciseRecord(exerciseSet.getExerciseRecord());
+        updated.setSetNumber(exerciseSet.getSetNumber());
+        updated.setActualRepetitions(exerciseSet.getActualRepetitions());
+        updated.setRepsOrDuration(exerciseSet.getRepsOrDuration());
+
+        return exerciseSetRepository.save(exerciseSet);
     }
 
     public void deleteExerciseSet(Long id, Long exerciseRecordId) {
+        if(!exerciseSetRepository.existsById(id) && !exerciseSetRepository.existsById(exerciseRecordId)) {
+            throw new ResourceNotFoundException("Exercise Set not found for id " + id + " and exercise record " + exerciseRecordId);
+        }
         Optional<ExerciseSet> exerciseSet = getExerciseSetByIdAndExerciseRecord(id, exerciseRecordId);
         exerciseSet.ifPresent(exerciseSetRepository::delete);
     }
 
     public void deleteAllExerciseSetsByExerciseRecord(Long exerciseRecordId) {
+        if (!exerciseSetRepository.existsById(exerciseRecordId)) {
+            throw new ResourceNotFoundException("Exercise Set not found for id " + exerciseRecordId);
+        }
         List<ExerciseSet> exerciseSets = getExerciseSetsByExerciseRecord(exerciseRecordId);
         exerciseSetRepository.deleteAll(exerciseSets);
     }
